fix: 전체Feed 조회 무한 스크롤 예약어 문제 수정.(feedRepository 예약어 'in' 삭제)

+++ src/main/java/com/hanghae/greenstep/feed/FeedController.java
@@ -18,7 +18,7 @@ public class FeedController {
     }
 
     @GetMapping("/feed")
-    public ResponseEntity<?> getFeed(int lastFeedId, HttpServletRequest request){
+    public ResponseEntity<?> getFeed(Long lastFeedId, HttpServletRequest request){
         return feedService.getFeed(lastFeedId, request);
     }
